TST: split re-usable utilities from the happi test
